Move service files to services directory
